requirements.txt
scikit-learn
streamlit
joblib
numpy<2
Pillow
